seed: implement network reachability pruning
This commit adds a new primary goroutine to the NetworkView struct, the
reachabilityPruner. The reachabilityPruner’s job is to s to maintain
the allnodes and reachableNodes map. Each time a new node is added,
we'll attempt to see if we can actually connect to it. If so, then
we'll add it to the reachableNodes. Every hour we'll examine the
reachableNodes map to ensure that all posted nodes are still reachable,
if not, we'll demote them.

With this change, we ensure that we’ll only return a node to a client
in response to a query iff, we’re able to reach it reliably as well.

This is only a start. Future follow up changes may add things like
ensuring the responses return are from distinct ASes, have geographic
diversity, etc, etc.