seed: add mechanism to allow clients to query for authoritative IP over A record
This commit adds a (kinda tacky) mechanism to allow clients to query
for the IP address of the authoritative name sever for the DNS seed.
With this in place, clients are able to query “soa. IN A” which will
return the IP address of the authoritative name server for the DNS
seed. Once clients have this IP address, then they’re able to directly
query the server over TCP. This serves as a fallback mechanism in the
case that the client’s DNS resolver rejects our large-ish SRV records.