seed: create new ChainView to be able to support DNS for multiple chains
In this commit, we add a new struct: ChainView. This is meant to wrap an
existing network view, yet allow multiple network views for distinct
chains to exist. We'll be using this to allow the main DNS server to
maintain a set of chain views so it can handle queries for distinct
chains and networks.