seed: fix goroutine leak and pile up by timing out dial attempts
In this commit, we fix a bug in the seed that would cause it to OOM
every few weeks due to goroutine build up. Before this commit, we used
the regular Dial package function which would block indefinitely in
certain cases. This would cause hundreds of thousands of goroutines to
build up, eventually causing a full OOM.

We fix this issue by giving up after 5 seconds if the remote peer
doesn't respond to our TCP dial attempt.